Shift a few env bits into scripts/bitbake
This attempts to separate the bits we *require* to run bitbake with oe-core
via the wrapper script, and which are independent of the build environment
(PSEUDO_DISABLED, PSEUDO_BUILD, BBFETCH2) from those which are more particular
to poky-init-build-env's way of setting things up (e.g. adding MACHINE to
BB_ENV_EXTRAWHITE, relying on OEROOT, etc).  This should make it easier to use
scripts/bitbake with non-standard workflows.

Signed-off-by: Chris Larson <chris_larson@mentor.com>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>