documentation/adt-manual/adt-prepare.xml: Added instruction for building ADT tarball
I added a note in the "Installing the ADT" section (2.1) saying that
if you need to build the ADT tarball you can use
'bitbake adt-installer'.  I also changed the location of the
toolchain from '...yocto-0.9' to '...yocto-1.0'.  Finally,
I changed the host sub-directory in the toolchain directory
from 'i586' to i686'.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>