elfutils_0.148.bb: Fix compilation on uclibc
Currently all components of elfutils dont compile with uclibc
but elfutils is one part which is needed by other recipes e.g.
gcc 4.5 to compile. we make adjustments so that when compiling
for uclibc targets it _only_ builds and packages libelf

use --enable-uclibc only when building for uclibc targets

The supporting patch is also needed for compiling with uclibc
to specify -lintl and -luargp on linker commandline

Add missing inherit on gettext

Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>