common-licenses: Adding new licenses and renaming
I'm adding full common licenses to the common license directory.

These licenses are taken from the SPDX project or the OSI license
repository. They conform to the SPDX naming convention and will be
used by the new license.bbclass license parser in order to maintain
some sort of internal standard for license naming.

Going forward, we should have a bigger conversation about the LICENSE
field and standardizing that to conform to this naming standard.

Signed-off-by: Beth Flanagan <elizabeth.flanagan@intel.com>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>