pseudo: Update pseudo to 1.1.1 version
Update both the core and pseudo_git packages to the latest 1.1.1 verison.

This fixes an issues where the call system() was not wrapped.  This could
lead to issues where certain spawned commands broke out of a pseudo-chroot
and created files in the wrong place.

Also the update the 1.0 -> 1.1.1 adds additional capabilities such as
beginning support for MacOS X.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>