kexec-tools: don't depend on virtual/kernel
There doesn't appear to be any terribly good reason for kexec-tools to
depend on the kernel.  (I verified that kexec-tools is buildable in a
clean TMPDIR without having previously built virtual/kernel.)

Having this dependency in place is a nuisance because it makes it
awkward to put kexec into an initramfs.  So, it seems like we would be
better off without.

Signed-off-by: Phil Blundell <philb@gnu.org>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>