linux-yocto: update meta SRCREV for new config groups
Updating the SRCREV for the kernel repo's meta branch to capture
the following commits:

  94fa015 meta: add taskstats experimental feature group
  4fb2ed5 meta: enable freezer support
  88d619e meta: enable fuse and cuse as modules
  f465827 meta: add namespaces + experimental configs
  fbdd376 meta: add devtmpfs config group
  b04f6d9 meta: re-enable cgroups options in the standard kernel

There's also a change to the recipe itself to trigger the taskstats
optional config items by default. This is to allow the introduction
of these changes gradually, since other recipes inheriting the kernel
can add or ignore these options at their convenience.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>