kernel: move menuconfig task after configure
Fixes [YOCTO 1136]

linux-yocto adds some configure steps that are necessary to prepare the source
tree after the do_patch task. This causes a "-c menuconfig" to fail in a clean
build tree. Typical use of menuconfig should be to modify the config provided
by the recipe being built. It therefor makes sense for the menuconfig task to
come after the configure task. This also happens to fix the issue seen with the
linux-yocto kernel recipe.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>