uclibc: Fix bug exposed by udev 168+ for mips architecture
newer udev uses sgnalfd + epoll and this exposed a bug
in uclibc for mips where SFD_NONBLOCK should be defined
with 0200 for mips but was using 04000. This would cause
random segfaults in udev during boot process

Tested on qemumips/angstrom console-image since angstrom
uses udev 171. It worked well when we did not use meta-oe
layer because udev in oe-core will then be used which is
at version 164.

Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>