image.bbclass: Add LINGUAS_INSTALL to dependency list
Currently if there are locales to install in the image the appropriate
packages might not automatically get built since there is no dependency
being exposed to bitbake.

This patch exposes the missing dependencies to bitbake through the RDEPENDS
field hence ensuring images build successfully when locales are added but
not depended upon anywhere else in the build.

[YOCTO #1215]

Signed-off-by: Saul Wold <sgw@linux.intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>