siteinfo.bbclass: Port over oe.dev logic for site files
In oe.dev we have a sets to pick out hostos/hostarch/etc site
files out of and include things like a common-linux site file.
This should also help out with adding multilib-specific site files
(ie x32).  In oe.dev we have an iterator but at Richard's request
we continue to return a list of files in siteinfo_get_files().

Signed-off-by: Tom Rini <tom_rini@mentor.com>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>