libpcre: Add bzip2, zlib and readline to DEPENDS
libpcre will build, if found, optional support for bzip2/zlib in
pcregrep and a readline frontend to pcretest.  It will not however
listen to --disable-pcregrep-{libz,libbz2} and will still build
these if found.  So we add bzip2, zlib and readline to DEPENDS to make
sure that we get consistent builds.

Signed-off-by: Tom Rini <tom_rini@mentor.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>